History API added.
Added a bit of javaScript to make it so when you try to go back after clicking on one of the buttons, it's the same as clicking on one of the previous buttons.

I wonder if the history API can be abused to lock users in your website until they close the tab. Good way to piss off users!